Separate the network config by family and add a "node_index" for each group.
This change allows a different network config for each family of hosts. For
instance, the controller may have a different network configuration than a
block storage node. This change adds a declaration for each family in the
overcloud-resource-registry-puppet.yaml. This change also adds the
"node_index" value to each group of servers. For instance, if there are 3
controllers, they will be assigned node_index of 0, 1, or 2. This will
allow static IP addresses to be assigned once this change to os-net-config
lands:
https://review.openstack.org/#/c/169937/

A use case for assigning a static IP is a provider network where there is
no DHCP server. This can be attached to each member of a family, and a
unique IP address will be assigned from the given subnet and host_ip_range.

Change-Id: I083df7ebbb535f97d8ddec2ac0e06281c55986cd

Conflicts:
	overcloud-without-mergepy.yaml
	puppet/ceph-storage-puppet.yaml